requirements.txt
websocket-client
fastapi
torch
requests
aiohttp
pytz
validators
pynvml
uvicorn